+++ src/Web/Sitemap.php
@@ -154,7 +154,7 @@ class Sitemap extends Component
     public function generateFileUrls(): array
     {
         $manager = Yii::$app->getUrlManager();
-        $defaultLanguages = $manager->hasI18nUrls() ? array_keys($manager->languages) : [null];
+        $defaultLanguages = $manager->i18nUrl ? array_keys($manager->languages) : [null];
         $urls = [];
 
         foreach ($this->views as $view) {