+++ tests/Modules/Admin/ModuleTest.php
@@ -6,15 +6,20 @@ namespace Hirtz\Skeleton\Tests\Modules\Admin;
 
 use Hirtz\Skeleton\Modules\Admin\Module;
 use Hirtz\Skeleton\Modules\Admin\ModuleInterface;
+use Hirtz\Skeleton\Models\User;
 use Hirtz\Skeleton\Test\TestCase;
+use Hirtz\Skeleton\Test\Traits\UserFixtureTrait;
 use Hirtz\Skeleton\Widgets\Navs\Nav;
 use Hirtz\Skeleton\Widgets\Navs\NavItem;
 use Hirtz\Skeleton\Widgets\Panels\Dashboard;
 use Hirtz\Skeleton\Widgets\Panels\DashboardItem;
+use ReflectionMethod;
 use Yii;
 
 class ModuleTest extends TestCase
 {
+    use UserFixtureTrait;
+
     protected Module $module;
 
     protected function setUp(): void
@@ -38,6 +43,49 @@ class ModuleTest extends TestCase
     {
         $nav = $this->module->dashboard(Dashboard::make());
         self::assertStringContainsString('Test Module', (string)$nav);
+    }
+
+    public function testSetLanguageFromIdentity(): void
+    {
+        User::updateAll(['language' => 'de'], ['id' => 1]);
+        Yii::$app->getUser()->login(User::findOne(1));
+
+        $this->setLanguage();
+
+        self::assertEquals('de', Yii::$app->language);
+    }
+
+    public function testSetIdentityLanguageFromRequest(): void
+    {
+        Yii::$app->getUser()->login(User::findOne(1));
+        Yii::$app->getRequest()->setQueryParams(['language' => 'de']);
+
+        $this->setLanguage();
+
+        self::assertEquals('de', User::findOne(1)->language);
+        self::assertEmpty($this->getLanguageCookieValue());
+    }
+
+    public function testSetGuestCookieLanguageFromRequest(): void
+    {
+        Yii::$app->getRequest()->setQueryParams(['language' => 'de']);
+
+        $this->setLanguage();
+
+        self::assertEquals('de', $this->getLanguageCookieValue());
+    }
+
+    protected function setLanguage(): void
+    {
+        Yii::$app->getI18n()->setLanguages(['en-US', 'de']);
+
+        $method = new ReflectionMethod($this->module, 'setLanguage');
+        $method->invoke($this->module, Yii::$app->getRequest());
+    }
+
+    protected function getLanguageCookieValue(): ?string
+    {
+        return Yii::$app->getResponse()->getCookies()->get(Yii::$app->getRequest()->languageParam)?->value;
     }
 }